Rename srcsets test helper
